+++ test_opensearchpy/test_async/test_server_secured/test_security_plugin.py
@@ -10,6 +10,7 @@
 
 from __future__ import unicode_literals
 
+import os
 from unittest import IsolatedAsyncioTestCase
 
 import pytest
@@ -42,8 +43,9 @@ class TestSecurityPlugin(IsolatedAsyncioTestCase):
 
     async def asyncSetUp(self) -> None:
         # pylint: disable=invalid-name, missing-function-docstring
+        password = os.environ.get("OPENSEARCH_INITIAL_ADMIN_PASSWORD", "admin")
         self.client = await get_test_client(
-            verify_certs=False, http_auth=("admin", "admin")
+            verify_certs=False, http_auth=("admin", password)
         )
         await add_connection("default", self.client)